Update Web App for control custom Telegram chat bot Update AdminBotApi.java Update CitiesService.java Update CitiesServiceImpl.java Update City.java Update TelegramBotApiListener.java

+++ src/main/java/com/resliv/bot/service/CitiesServiceImpl.java
@@ -4,11 +4,14 @@ import com.resliv.bot.entity.City;
 import com.resliv.bot.repository.CitiesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
 @Service
+@Transactional
 public class CitiesServiceImpl implements CitiesService {
 
     @Autowired
@@ -25,47 +28,66 @@ public class CitiesServiceImpl implements CitiesService {
     @Override
     public void createCity(String name) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
         }
-        City city = new City(0L, name, null);
-        citiesRepository.save(city);
-    }
-
-    // Один арг. descr.
-    @Override
-    public void createCity(String name, String... description) {
-        if (name == null) {
-            throw new IllegalArgumentException("City is null!");
-        }
-        City city = new City(0L, name, Arrays.asList(description.clone()));
+        City city = new City(0L, name, Collections.emptyList());
         citiesRepository.save(city);
     }
 
     @Override
-    public void updateCityByNameNewDescription(String name, String... description) {
+    public void createCity(String name, String description) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
+        } else if (description == null) {
+            throw new IllegalArgumentException("City description is null!");
+        }
+        List<String> descriptions = new ArrayList<>();
+        descriptions.add(description);
+        City city = new City(0L, name, descriptions);
+        citiesRepository.save(city);
+    }
+
+    @Override
+    public void updateCityByNameNewDescription(String name, String... descriptions) {
+        if (name == null) {
+            throw new IllegalArgumentException("City name is null!");
+        }
+        if (!citiesRepository.existsByName(name)) {
+            throw new RuntimeException("City with such name does not exist");
         }
         City city = citiesRepository.findByName(name);
-        city.setDescriptions(Arrays.asList(description.clone()));
+        city.setDescriptions(Arrays.asList(descriptions.clone()));
+        // Выдает ошибку
         citiesRepository.save(city);
     }
 
     @Override
-    public void updateCityByNameAddDescription(String name, String... description) {
+    public void updateCityByNameAddDescription(String name, String... descriptions) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
+        }
+        if (!citiesRepository.existsByName(name)) {
+            throw new RuntimeException("City with such name does not exist");
         }
         City city = citiesRepository.findByName(name);
-        List<String> extraDescriptions = Arrays.asList(description.clone());
-        city.getDescriptions().addAll(extraDescriptions);
+        List<String> extraList = Arrays.asList(descriptions.clone());
+        List<String> currentList;
+        if ((currentList = city.getDescriptions()) == null) {
+            city.setDescriptions(extraList);
+        } else {
+            currentList.addAll(extraList);
+            city.setDescriptions(currentList);
+        }
         citiesRepository.save(city);
     }
 
     @Override
     public void deleteCityByName(String name) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
+        }
+        if (!citiesRepository.existsByName(name)) {
+            throw new RuntimeException("City with such name does not exist");
         }
         citiesRepository.deleteByName(name);
     }
@@ -75,7 +97,7 @@ public class CitiesServiceImpl implements CitiesService {
     @Override
     public boolean existCityByName(String name) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
         }
         return citiesRepository.existsByName(name);
     }
@@ -83,7 +105,7 @@ public class CitiesServiceImpl implements CitiesService {
     @Override
     public List<String> getDescriptionsByCityName(String name) {
         if (name == null) {
-            throw new IllegalArgumentException("City is null!");
+            throw new IllegalArgumentException("City name is null!");
         }
         City city = citiesRepository.findByName(name);
         if (city != null) {
